Fix: Updated package to io.github.farhazulmullick.lenslogger and module to lens-logger.
