+++ composeApp/src/commonMain/kotlin/io/github/farhazulmullick/network/HttpClient.kt
@@ -2,7 +2,7 @@ package io.github.farhazulmullick.network
 
 import io.github.aakira.napier.DebugAntilog
 import io.github.aakira.napier.Napier
-import io.github.farhazulmullick.lensktor.plugin.network.LensHttpLogger
+import io.github.farhazulmullick.lenslogger.plugin.network.LensHttpLogger
 import io.ktor.client.HttpClient
 import io.ktor.client.HttpClientConfig
 import io.ktor.client.plugins.DefaultRequest
