+++ lens-logger/build.gradle.kts
@@ -1,0 +1,114 @@
+plugins {
+    alias(libs.plugins.kotlinMultiplatform)
+    alias(libs.plugins.androidKotlinMultiplatformLibrary)
+    alias(libs.plugins.composeMultiplatform)
+    alias(libs.plugins.composeCompiler)
+    alias(libs.plugins.kotlinSerialization)
+    alias(libs.plugins.vanniktech.publish)
+}
+
+mavenPublishing {
+    coordinates(
+        groupId = "io.github.farhazulmullick",
+        artifactId = "lens-logger",
+        version = libs.versions.lens.get()
+    )
+
+    pom {
+        name.set("Lens Ktor")
+        description.set("A description of what my library does.")
+        inceptionYear.set("2025")
+        url.set("https://github.com/farhazulMullick/Lens/")
+        licenses {
+            license {
+                name.set("MIT License")
+                url.set("https://opensource.org/licenses/MIT")
+                distribution.set("repo")
+            }
+        }
+        developers {
+            developer {
+                id.set("farhazulMullick")
+                name.set("Farhazul Mullick")
+                url.set("https://github.com/farhazulMullick/")
+            }
+        }
+        scm {
+            url.set("https://github.com/farhazulMullick/Lens/")
+            connection.set("scm:git:git://github.com/farhazulMullick/Lens.git")
+            developerConnection.set("scm:git:ssh://git@github.com/farhazulMullick/Lens.git")
+        }
+    }
+}
+
+kotlin {
+    androidLibrary {
+        namespace = "io.github.farhazulmullick.lensloggger"
+        compileSdk = 35
+        minSdk = 24
+    }
+    val xcfName = "lensktorKit"
+
+    iosX64 {
+        binaries.framework {
+            baseName = xcfName
+        }
+    }
+
+    iosArm64 {
+        binaries.framework {
+            baseName = xcfName
+        }
+    }
+
+    iosSimulatorArm64 {
+        binaries.framework {
+            baseName = xcfName
+        }
+    }
+
+// Source set declarations.
+// Declaring a target automatically creates a source set with the same name. By default, the
+// Kotlin Gradle Plugin creates additional source sets that depend on each other, since it is
+// common to share sources between related targets.
+// See: https://kotlinlang.org/docs/multiplatform-hierarchy.html
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation(libs.kotlin.stdlib)
+                implementation(compose.material3)
+                implementation(compose.runtime)
+                implementation(compose.foundation)
+                implementation(compose.ui)
+                implementation(compose.materialIconsExtended)
+                implementation(compose.components.resources)
+                implementation(compose.components.uiToolingPreview)
+                implementation(libs.androidx.lifecycle.viewmodel)
+                implementation(libs.androidx.lifecycle.runtimeCompose)
+                implementation(libs.androidx.navigation)
+
+                implementation(libs.ktor.client.core)
+                implementation(libs.ktor.client.logging)
+                implementation(libs.ktor.client.content.negotiation)
+                implementation(libs.ktor.client.serializer)
+                implementation(libs.ktor.json.serializer)
+                implementation(libs.napier)
+            }
+        }
+
+        commonTest {
+            dependencies {
+                implementation(libs.kotlin.test)
+            }
+        }
+
+        androidMain {
+            dependencies {}
+        }
+
+        iosMain {
+            dependencies {}
+        }
+    }
+
+}
