+++ lens-logger/src/androidMain/AndroidManifest.xml
