+++ lens-logger/src/androidMain/kotlin/io/github/farhazulmullick/lenslogger/Platform.android.kt
@@ -1,0 +1,3 @@
+package io.github.farhazulmullick.lenslogger
+
+actual fun platform() = "Android"
