+++ lens-logger/src/commonMain/kotlin/io/github/farhazulmullick/lenslogger/Platform.kt
@@ -1,0 +1,3 @@
+package io.github.farhazulmullick.lenslogger
+
+expect fun platform(): String
