+++ lens-logger/src/commonMain/kotlin/io/github/farhazulmullick/lenslogger/navigation/LensRoute.kt
@@ -1,0 +1,12 @@
+package io.github.farhazulmullick.lenslogger.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed class LensRoute {
+    @Serializable
+    object NetLogScreen: LensRoute()
+
+    @Serializable
+    data class NetLogInfoScreen(val index: Int): LensRoute()
+}
