+++ lens-logger/src/iosMain/kotlin/io/github/farhazulmullick/lenslogger/Platform.ios.kt
@@ -1,0 +1,3 @@
+package io.github.farhazulmullick.lenslogger
+
+actual fun platform() = "iOS"
