--- lens-module/src/androidMain/kotlin/io/github/farhazulmullick/lensktor/Platform.android.kt
@@ -1,3 +1,0 @@
-package io.github.farhazulmullick.lensktor
-
-actual fun platform() = "Android"
