--- lens-module/src/commonMain/kotlin/io/github/farhazulmullick/lensktor/Platform.kt
@@ -1,3 +1,0 @@
-package io.github.farhazulmullick.lensktor
-
-expect fun platform(): String
