--- lens-module/src/commonMain/kotlin/io/github/farhazulmullick/lensktor/modal/Resource.kt
@@ -1,22 +1,0 @@
-package io.github.farhazulmullick.lensktor.modal
-
-sealed class Resource<T> {
-
-    /**
-     * Operation is currently loading.
-     */
-    class Loading<Domain> : Resource<Domain>()
-
-    /**
-     * Successful operation with data.
-     */
-    class Success<Domain>(val data: Domain) : Resource<Domain>()
-
-    /**
-     * Operation resulted in an error.
-     */
-    data class Failed<Domain>(
-        val data: Domain? = null,
-        val throwable: Throwable? = null
-    ) : Resource<Domain>()
-}
