--- lens-module/src/commonMain/kotlin/io/github/farhazulmullick/lensktor/navigation/LensNavHost.kt
@@ -1,2 +1,0 @@
-package io.github.farhazulmullick.lensktor.navigation
-
