--- lens-module/src/commonMain/kotlin/io/github/farhazulmullick/lensktor/navigation/LensRoute.kt
@@ -1,12 +1,0 @@
-package io.github.farhazulmullick.lensktor.navigation
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-sealed class LensRoute {
-    @Serializable
-    object NetLogScreen: LensRoute()
-
-    @Serializable
-    data class NetLogInfoScreen(val index: Int): LensRoute()
-}
