--- lens-module/src/iosMain/kotlin/io/github/farhazulmullick/lensktor/Platform.ios.kt
@@ -1,3 +1,0 @@
-package io.github.farhazulmullick.lensktor
-
-actual fun platform() = "iOS"
